Conflicts:
	README.md

+++ app/controllers/products_controller.rb
@@ -1,9 +1,17 @@
 class ProductsController < ApplicationController
+before_action :author!, only: [:edit, :update]
   expose(:category)
   expose(:products)
   expose(:product)
   expose(:review) { Review.new }
   expose_decorated(:reviews, ancestor: :product)
+
+def author!
+  unless self.product.user == current_user
+    redirect_to category_product_url(category, product),
+      flash: { error: 'You are not allowed to edit this product.' }
+  end
+end
 
   def index
   end
@@ -15,31 +23,60 @@ class ProductsController < ApplicationController
   end
 
   def edit
+
   end
 
   def create
-    self.product = Product.new(product_params)
+if user_signed_in?	
 
+
+
+    self.product = Product.new(product_params)
+    self.product.user_id = current_user.id
     if product.save
       category.products << product
       redirect_to category_product_url(category, product), notice: 'Product was successfully created.'
     else
       render action: 'new'
     end
+else
+redirect_to new_user_session_url
+
+end
+
   end
 
   def update
+if user_signed_in?	
+
+
+
     if self.product.update(product_params)
       redirect_to category_product_url(category, product), notice: 'Product was successfully updated.'
     else
       render action: 'edit'
     end
+
+
+
+
+else
+redirect_to new_user_session_url
+
+end
   end
 
   # DELETE /products/1
   def destroy
+if user_signed_in?
     product.destroy
     redirect_to category_url(product.category), notice: 'Product was successfully destroyed.'
+
+else
+redirect_to new_user_session_url
+
+end
+
   end
 
   private
